+++ src/cli/commands/view.rs
@@ -94,7 +94,7 @@ impl Command {
     );
 
     if !output.is_empty() {
-      pager::output(&output, &ctx.config, self.pager)?;
+      pager::output(&output, &ctx.config, self.pager || ctx.use_pager)?;
     }
 
     Ok(())